adding css

+++ client/src/components/login.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
-import {login, userPublicPosts,userPrivatePosts} from "../utils/API";
-// import './loginSignup.css';
+import {login} from "../utils/API";
+import './loginSignup.css';
 
 export default class Login extends Component {
 
@@ -9,45 +9,49 @@ export default class Login extends Component {
         password: "" 
     }
 
-    getPublicPosts = (username) => {
-        userPublicPosts(username)
-        .then(res => {
-          localStorage.setItem("userPublicPost", JSON.stringify(res.data));
-        })
-      }
-    
-    getPrivatePosts = (username) => {
-        userPrivatePosts(username)
-        .then(res => {
-          localStorage.setItem("userPrivatePost", JSON.stringify(res.data));
-        })
-      }
-
     handleSignIn = () => {
-        console.log("Sending user info for sign in");
-        login(this.state)
-        .then(res => {
-            if (res.data === null) {
-                localStorage.clear();
-                alert("Username or Password is incorrect!");
-                this.setState({email: "", password: ""});
-            }
-            else {
-                localStorage.setItem("currentUser", JSON.stringify(res.data));
-                // this.getPublicPosts(this.state.email);
-                // this.getPrivatePosts(this.state.email);
-                this.props.setUser(res.data);
-
-                window.location.replace("/member")
-            }
-        });
+        if (this.state.email && this.state.password) {
+            login(this.state)
+            .then(res => {
+                if (res.data === null) {
+                    localStorage.clear();
+                    alert("Username or Password is incorrect!");
+                    this.setState({email: "", password: ""});
+                }
+                else {
+                    localStorage.setItem("currentUser", JSON.stringify(res.data))
+                    this.props.setUser(res.data);
+                    window.location.replace("/member")
+                }
+            });
+        }
+        else {
+            alert("Please fill up all the fields");
+            this.setState({email: "", password: ""});
+        }
     }
-
-    
 
     handleChange = ({target:{name,value}}) => this.setState({[name]:value})
     render() {
         return (
+            <div className="App">
+            {!this.state.user ? <nav className="navbar navbar-expand-lg navbar-light fixed-top">
+              <div className="container">
+                <a href="/" class="logo">SAFESPACE</a>
+                <div className="collapse navbar-collapse" id="navbarTogglerDemo02">
+                  <ul className="navbar-nav ml-auto">
+                    <li className="nav-item">
+                    <a href="/" class="login">Login</a>
+                    </li>
+                    <li className="nav-item">
+                    <a href="/sign-up" class="signup">Sign up</a>
+                    </li>
+                  </ul>
+                </div>
+              </div>
+            </nav> : <button onClick={this.handleLogout}>Log Out</button>}
+            <div className="auth-wrapper">
+              <div className="auth-inner">
             <form>
                 <h3>Sign In</h3>
 
@@ -73,6 +77,9 @@ export default class Login extends Component {
                     Forgot <a href="#">password?</a>
                 </p>
             </form>
+            </div>
+            </div>
+            </div>
         );
     }
 }
